% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AA3to1.R
\name{AA3to1}
\alias{AA3to1}
\title{Convert amino acids from 3 letter to 1 letter}
\usage{
AA3to1(AA)
}
\arguments{
\item{AA}{a string of 3 letter AA}
}
\value{
A string of 1 letter AA
}
\description{
Convert amino acids from 3 letter to 1 letter
}
