% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignNonOverlapGroup.R
\name{AssignNonOverlapGroup}
\alias{AssignNonOverlapGroup}
\title{Assign groups with minimal overlaps}
\usage{
AssignNonOverlapGroup(start, end, min_dist = 0)
}
\arguments{
\item{start}{start positions}

\item{end}{end positions}

\item{min_dist}{minimal distance to merge two groups}
}
\value{
a vector of the group ids
}
\description{
Group genomic features into non overlapping groups. Each entry/row is one feature.
Start and end describe the start end positions of the features. Features with in each group have to
be at least min_dist away from each other.
}
