% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Variants_AlphaFold.R
\name{Color_Variants_AlphaFold}
\alias{Color_Variants_AlphaFold}
\title{Map variants and ET scores to AlphaFold structures}
\usage{
Color_Variants_AlphaFold(
  variants_case,
  variants_ctrl = NULL,
  prot_id,
  color_type = c("ET", "red_white", "red_white_blue", "gray_scale", "rev_EA_bin",
    "red_yellow_green"),
  additional_pymol_cmds = NULL,
  pml_output,
  output_format = c("pml", "pse")
)
}
\arguments{
\item{variants_case}{A dataframe that stores the mutations in the cases in target gene.
The dataframe should contain at least 2 columns: SUB and EA. Single letter animo acid code should
be used in SUB, eg. L10P. EA can be replaced with other values, but have to scale to the
range of (0, 100].}

\item{variants_ctrl}{A dataframe that stores the mutations in the controls in target gene.}

\item{prot_id}{Protein identification. Use full ENSP for human proteins, eg. ENSP00000388638.1.
Use locus tag for E. coli MG1655 proteins, eg. b4112.}

\item{color_type}{The type of color range to use. Available selections: "ET",
"red_white", "red_white_blue", "gray_scale", "rev_EA_bin", "red_yellow_green". The colors
will be used for ET. EA will be colored using the reverse of this coloring.}

\item{additional_pymol_cmds}{Additional pymol commands to append to the pymol script. Store them in a vector.}

\item{pml_output}{Path to store the pymol command file.}

\item{output_format}{Output pymol script (pml) or pymol session (pse). Pse output requires pymol to
be accessble from terminal with "pymol".}
}
\value{
A tibble that contains the coloring information. The pml file will
be written to the location indicated by pml_output.
}
\description{
Map variants and ET scores to AlphaFold structures. Pymol is required to view the pml
file. Network connection is also required to load the pdb file from AlphaFold.
}
