% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPDBByET.R
\name{ColorPDBByET}
\alias{ColorPDBByET}
\title{Pymol color pdb structure using ET}
\usage{
ColorPDBByET(
  pdb_file,
  chain,
  ET_format = c("EA", "UET", "ENSP"),
  ET_file,
  color_type = c("ET", "red_white", "red_white_blue", "gray_scale"),
  coverage_cutoff = 1,
  output_file,
  output_format = c("pml", "pse")
)
}
\arguments{
\item{pdb_file}{Path to pdb file. Can be path to local file or url ("https://files.rcsb.org/view/1xb7.pdb").}

\item{chain}{The chain in the pymol session that needs to be colored.}

\item{ET_format}{Specify the ET file format. "UET" for ET server, "EA" for ET in EA output, and "ENSP"
for precalculated ET scores from EA VEP output.}

\item{ET_file}{Path to ET file. If "ENSP" is used in ET_format, provide ENSP id here, including version.}

\item{color_type}{The type of color range to use. Available selections: "ET",
"red_white", "red_white_blue", and "gray_scale".}

\item{coverage_cutoff}{Value between (0,1]. Only residue below this cutoff will be colored.}

\item{output_file}{Path to write the output pymol file.}

\item{output_format}{Output pymol script (pml) or pymol session (pse). Pse output requires pymol to
be accessible from terminal with "pymol".}

\item{pml_output}{Path to store the pymol command file.}
}
\value{
A tibble with the sequence alignment between the linear sequence in ET file and the sequence in
the target chain in PDB.
}
\description{
The function first align the linear sequence from the ET file to the sequence in PDB. PDB
structure is then colored according the ET scores. A pymol output file is written to provided path.
The function returns an alignment table. The percent matching between sequence in the ET file and PDB
is printed in the console.
}
\examples{
ColorPDBByET(pdb_file = "https://files.rcsb.org/view/2pjl.pdb", chain = "A", ET_format = "ENSP",
             ET_file = "ENSP00000000442.6", color_type = "ET", coverage_cutoff = 0.3,
             output_format = "pml", output_file = "test.pml")
}
