% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareLinearSeqs.R
\name{CompareLinearSeqs}
\alias{CompareLinearSeqs}
\title{Compare two protein linear sequences}
\usage{
CompareLinearSeqs(
  seq1,
  seq2,
  pos.only = TRUE,
  penalty = c("blastp", "pyETv", "DNA")
)
}
\arguments{
\item{seq1}{Character string. Linear sequence of protein 1}

\item{seq2}{Character string. Linear sequence of protein 2}

\item{pos.only}{If TRUE, then only matched positions between the pdb sequence and
the linear sequence are returned.}

\item{penalty}{Gap penalties used for the pairwise alignment.}
}
\value{
A tibble that contains the matching information of the pdb sequence and
linear sequence.
}
\description{
Compared two linear sequences. The matching positions in the two
linear sequences are returned.
}
