% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareSeqs.R
\name{CompareSeqs}
\alias{CompareSeqs}
\title{Compare protein sequence in a pdb file with linear seq}
\usage{
CompareSeqs(
  pdb_file,
  chain,
  seq,
  pos.only = TRUE,
  penalty = c("blastp", "pyETv"),
  pdb_format = c("pdb", "pdbx")
)
}
\arguments{
\item{pdb_file}{path to PDB file}

\item{chain}{Character string. Chains of interest}

\item{seq}{Character string. Linear sequence of the protein}

\item{pos.only}{If TRUE, then only matched positions between the pdb sequence and
the linear sequence are returned.}

\item{penalty}{Gap penalties used for the pairwise alignment.}

\item{pdb_format}{Use "pdb" if the input is traditional pdb. Use "pdbx" for PDBx/mmCIF.}
}
\value{
A tibble that contains the matching information of the pdb sequence and
linear sequence.
}
\description{
Sequence from pdb_file is extracted and compared (aligned) to linear
sequence. The matching positions in the pdb entry and in the linear sequence are
returned.
}
