% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCW_z.R
\name{ComputeSCWzscore}
\alias{ComputeSCWzscore}
\title{Compute SCW z scores}
\usage{
ComputeSCWzscore(SCW_background, resi, output_df = FALSE)
}
\arguments{
\item{SCW_background}{A list with parameters required for SCW z score calculation.
This is obtained through GetSCWBackground function.}

\item{resi}{numeric vector. The selection of residue positions. Residue positions
should be based on pdb structure not the sequence positions}

\item{output_df}{Whether a df should be returned instead of a vector.}
}
\value{
a vector or df with biased and unbiased SCW z scores.
}
\description{
This function calculates the biased and unbiased SCW z scores
for a given selection of residues in the structure.
}
