% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCW_z.R
\name{ComputeSCWzscore_list}
\alias{ComputeSCWzscore_list}
\title{Compute SCW z scores with multiple selection of residues}
\usage{
ComputeSCWzscore_list(SCW_background, resi_list)
}
\arguments{
\item{SCW_background}{A list with parameters required for SCW z score calculation.
This is obtained through GetSCWBackground function.}

\item{resi_list}{a list of numeric vector. The selection of residue positions.
Residue positions should be based on pdb structure not the sequence positions.}
}
\value{
df with biased and unbiased SCW z scores for each residue selection in
the resi_list. Names of the resi_list is used as the condition column. If no names
are provided, list numbering is used.
}
\description{
This function calculates the biased and unbiased SCW z scores
for several selections of residues in the structure.
}
