% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadPDB.R
\name{DownloadPDB}
\alias{DownloadPDB}
\title{Rewrite PDB file according to linear sequence}
\usage{
DownloadPDB(pdb, out_dir = ".")
}
\arguments{
\item{pdb}{The pdb id that needs to be downloaded}

\item{out_dir}{The directory that the pdb will be saved}
}
\value{
a message reporting the pdb is downloaded
}
\description{
This function intake a pdb id and download the pdb file to
the path location as {pdb.id}.pdb
}
