% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchET.R
\name{FetchET}
\alias{FetchET}
\title{Fetch prestored ET file}
\usage{
FetchET(prot_id)
}
\arguments{
\item{prot_id}{Protein id. Protein identification. Use full ENSP for human proteins, eg. ENSP00000388638.1.
Use locus tag for E. coli MG1655 proteins, eg. b4112.}
}
\value{
A dataframe(tibble) with prestored ET. The ET file is reduced to only include POS, AA and coverage.
}
\description{
Fetch prestored ET file. Only supports human and E. coli MG1655 proteins
}
