% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCoordinates.R
\name{GetCoordinates_no_filter}
\alias{GetCoordinates_no_filter}
\title{Extract coordinates from PDB file without AA filter}
\usage{
GetCoordinates_no_filter(pdb_file, chain)
}
\arguments{
\item{pdb_file}{path to PDB file}

\item{chain}{Character string. Chains that needs to be extracted}
}
\value{
A tibble that contains chain, POS (AA position), AA, ATOM, ATOM_id, x, y, z
}
\description{
Extract coordinates from PDB file. Output will be a tibble
that contains protein chain, AA position, AA type (one letter), atom type,
atom id and coordinates (x, y, z)
}
