% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCoordinates.R
\name{GetCoordinates_pdb}
\alias{GetCoordinates_pdb}
\title{Extract coordinates from PDB file}
\usage{
GetCoordinates_pdb(pdb_file, chain, CA_only = TRUE, remove_insertions = TRUE)
}
\arguments{
\item{pdb_file}{path to PDB file}

\item{chain}{Character string. Chains that needs to be extracted}

\item{CA_only}{Logic. Whether only C alpha are extracted}

\item{remove_insertions}{Logic. Some residues are labeled as insertions in the pdb
file and they have the same POS index as the residue that they insert after. If True,
these insertion residues are removed.}
}
\value{
A tibble that contains chain, POS (AA position), AA, ATOM, ATOM_id, x, y, z
}
\description{
Extract coordinates from PDB file. Output will be a tibble
that contains protein chain, AA position, AA type (one letter), atom type,
atom id and coordinates (x, y, z)
}
