% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDistanceMatrix.R
\name{GetDistanceMatrix}
\alias{GetDistanceMatrix}
\title{Get residue distance matrix/dataframe}
\usage{
GetDistanceMatrix(df, output_type = c("df", "matrix"))
}
\arguments{
\item{df}{atom coordinates dataframe from GetCoordinates function.
It should contain these columns POS, ATOM, x, y, and z}

\item{output_type}{output as dataframe/tibble or matrix}
}
\value{
a tibble with POS_i, POS_j and distance or a distance matrix
}
\description{
Calculates distance matrix for residues from atom distances.
Distance between two residues is defined as the minimal distance
between any two atoms within those residues.
To calculate distance with only C-alpha, only include C-alpha atoms
in the input df.
}
