% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInterface.R
\name{GetInterface}
\alias{GetInterface}
\title{Get residues at protein interface of two chains from a PDB file}
\usage{
GetInterface(
  pdb_file,
  chain1,
  chain2,
  dist_cutoff = 4,
  output_type = c("paired", "stacked")
)
}
\arguments{
\item{pdb_file}{path to pdb file}

\item{chain1}{the first chain}

\item{chain2}{the second chain}

\item{dist_cutoff}{the distance cutoff that consider two atoms are adjacent}

\item{output_type}{"paired" or "stacked"}
}
\value{
a dataframe(tibble) contains all residues at the protein interface
}
\description{
Residues are at the interface when any of their heavy atoms are
within the cutoff range of a heavy atom from the other chain. The paired output
contains columns chain_i, chain_j, POS_i, POS_j, and dist. The stacked output
contains columns chain and POS.
}
