% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLaplacianMatrix.R
\name{GetLaplacianMatrix}
\alias{GetLaplacianMatrix}
\title{Get Laplacian Matrix for a given distance matrix}
\usage{
GetLaplacianMatrix(dist_mat, cutoff = 4, normalize = FALSE)
}
\arguments{
\item{dist_mat}{Distance matrix. Can be obtained from GetDistanceMatrix.}

\item{cutoff}{dbl. The maximal distance two residues are consider adjacent.}

\item{normalize}{logic. whether normalized Laplacian is returned}
}
\value{
The Laplacian matrix
}
\description{
Compute Laplacian Matrix from given distance matrix. Distance matrix
can be obtained from GetDistanceMatrix function.
}
