% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLaplacianSmoothness.R
\name{GetLaplacianSmoothness}
\alias{GetLaplacianSmoothness}
\title{Compute smoothness of residue ranks over a given structure.}
\usage{
GetLaplacianSmoothness(
  dist_mat,
  vari,
  cutoff = 4,
  variable_normalization = c("L2", "L1", "none"),
  seed = 100
)
}
\arguments{
\item{dist_mat}{Distance matrix. Can be obtained from GetDistanceMatrix.}

\item{vari}{a numerical vector that stores the ranks/scores of the residues. The
length of vari should be the same as the dim of dist_mat. The order should also be
matched.}

\item{cutoff}{dbl. The maximal distance two residues are consider adjacent.}

\item{variable_normalization}{Normallization type of the variable. Default is L2.}

\item{seed}{dbl. Seeding for randomization.}
}
\value{
A list that stores the smoothness of the vari vector, 10000 random smoothness,
the z score and p value.
}
\description{
Compute Laplacian smoothness of vector vari over the protein structure.
Smoothness background is computed for 10000 random vari vector which are generated
by shuffling the vari vector. Z score and p values (normal, one sided) are calculated
for the vari smoothness comparing to background.
}
