% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetpLDDT.R
\name{GetpLDDT}
\alias{GetpLDDT}
\title{Get per-residue confidence metric (pLDDT) from AlphaFold structures}
\usage{
GetpLDDT(pdb_file, chain = "A")
}
\arguments{
\item{pdb_file}{path to AlphaFold PDB file}

\item{chain}{Character vector. Chains of interest. Usually AlphaFold structures are stored
in chain A.}
}
\value{
A tibble that contains the chains, amino acid positions and pLDDT scores.
}
\description{
Amino acid positions and pLDDT scores are extracted from AlphaFold pdb
file. pLDDT values are stored in b factor column. So this function can also be used
to extract b factor from experimental pdb files. Use cation when applying to experimental
data due to isoforms have the same AA position.
}
