% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR_filter.R
\name{IQR_filter}
\alias{IQR_filter}
\title{Apply 1.5 IQR filter for a given numeric vector}
\usage{
IQR_filter(x, type = c("bool", "ends"))
}
\arguments{
\item{x}{A given numeric vector to apply the filter}

\item{type}{If "bool" is used, them a boolean vector with the same length as x is returned.
If "ends" is used, then the lower and upper boundaries are returned}
}
\value{
A boolean vector or the lower and upper boundaries
}
\description{
Apply IQR to vector x.
}
