% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LollipopPlot.R
\name{LollipopPlot}
\alias{LollipopPlot}
\title{Plot lollipop plot}
\usage{
LollipopPlot(
  variants,
  prot_id,
  prot_color = "ET",
  plot_domain = TRUE,
  AC_scale = c("log", "linear"),
  show_EA_bin = TRUE,
  EA_color = c("prismatic", "gray_scale", "EA_bin", "black"),
  fix_scale = TRUE,
  domain_min_dist = 0,
  pad_ratio = 0.1,
  title = NULL,
  add_legend = TRUE
)
}
\arguments{
\item{variants}{A dataframe with the variants in the cases. "SUB", "EA" and "AC" columns
are required. SUB should be 1 letter format, e.g. S10L. EA should be numerical between 0-100.
AC is allele count.}

\item{prot_id}{Protein identification. Use ENSP id for human proteins and b number for E. coli
proteins.}

\item{prot_color}{Linear protein color. Default set as ET.}

\item{plot_domain}{logical. When domain annotations are available, whether plot domain
track.}

\item{AC_scale}{Whether to use linear or log scale for allele count.}

\item{show_EA_bin}{When TRUE, the circle sizes in the lollipop plot are adjusted based on
EA bins. High: (70,100], medium: (30,70] and low (0,30].}

\item{EA_color}{Coloring style for EA (circle in the lollipop plots). "prismatic": higher EA
mutations are warmer. "gray_scale": higher EA mutations are darker. "EA_bin": based on 3 EA bins, high,
medium and low. "black": uniform black color.}

\item{domain_min_dist}{Controls the minimum distances between two domains in the domain track.
If the distance between two domains are less than domain_min_dist, they will be plotted in separate lines.
If domain annotations overlap, set to a larger value.}

\item{pad_ratio}{Controls the extra space between the ET (center) track and the mutation tracks.
Use larger values for larger space.}

\item{title}{Title for the plot.}

\item{add_legend}{Whether to include ET/EA legends at the bottom of the plot.}
}
\value{
lollipop plot
}
\description{
This function graphs a lollipop plot to show mutational profile in a given gene. The center ET track is colored as prismatic style, with the most important
ET positions as red. The height of the lollipops reflects allele count. The color and/or size of the
circles reflects EA scores for the mutations. This function fetch pre stored ET scores, which only
works for human and E. coli reference proteins. See LollipopPlot2 for ploting mutations in both case and
control.
}
\examples{
# Prepare variant data
set.seed(755)
mut_case <- read_tsv(file.path(system.file("extdata", package = "EvoTrace"),
                              "basS_muts.tsv"),
                     show_col_types = FALSE) \%>\%
  mutate(AC = runif(11, 0, 1000)) \%>\%
  mutate(AC = round(AC)) \%>\%
  mutate(AF = AC/1000)

# Make log scale plot using prismatic coloring
LollipopPlot(variants = mut_case, prot_id = "b4112",
             AC_scale = "log", plot_domain = TRUE,
             show_EA_bin = TRUE, EA_color = "prismatic")

# Make linear scale plot using EA_bin coloring
LollipopPlot(variants = mut_case, prot_id = "b4112",
             AC_scale = "linear", plot_domain = FALSE,
             show_EA_bin = TRUE, EA_color = "EA_bin",
             pad_ratio = 0)
}
