% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LollipopPlot2.R
\name{LollipopPlot2}
\alias{LollipopPlot2}
\title{Plot lollipop plot for case and control}
\usage{
LollipopPlot2(
  variants_case,
  variants_ctrl,
  prot_id,
  prot_color = "ET",
  plot_domain = TRUE,
  AC_scale = c("log", "linear"),
  show_EA_bin = TRUE,
  EA_color = c("prismatic", "gray_scale", "EA_bin", "black"),
  fix_scale = TRUE,
  pad_ratio = 0.05,
  domain_min_dist = 0,
  title = NULL,
  y_lab = "Allele Count",
  add_legend = TRUE,
  AA_start = NA,
  AA_end = NA,
  return_individual_plots = FALSE
)
}
\arguments{
\item{variants_case}{A dataframe with the variants in the cases. "SUB", "EA" and "AC" columns
are required. SUB should be 1 letter format, e.g. S10L. EA should be numerical between 0-100.
AC is allele count.}

\item{variants_ctrl}{A dataframe with the variants in the controls. Same requirements as
variants_case.}

\item{prot_id}{Protein identification. Use ENSP id for human proteins and b number for E. coli
proteins.}

\item{prot_color}{Linear protein color. Default set as ET.}

\item{plot_domain}{logical. When domain annotations are available, whether plot domain
track.}

\item{AC_scale}{Whether to use linear or log scale for allele count.}

\item{show_EA_bin}{When TRUE, the circle sizes in the lollipop plot are adjusted based on
EA bins. High: (70,100], medium: (30,70] and low (0,30].}

\item{EA_color}{Coloring style for EA (circle in the lollipop plots). "prismatic": higher EA
mutations are warmer. "gray_scale": higher EA mutations are darker. "EA_bin": based on 3 EA bins, high,
medium and low. "black": uniform black color.}

\item{fix_scale}{Whether axis for AC in case and control should be the same scale or not.}

\item{pad_ratio}{Controls the extra space between the ET (center) track and the mutation tracks.
Use larger values for larger space.}

\item{domain_min_dist}{Controls the minimum distances between two domains in the domain track.
If the distance between two domains are less than domain_min_dist, they will be plotted in separate lines.
If domain annotations overlap, set to a larger value.}

\item{title}{Title for the plot.}

\item{y_lab}{Label for y axis. Default "Allele Count".}

\item{add_legend}{Whether to include ET/EA legends at the bottom of the plot.}

\item{AA_start}{The starting residue position for the lollipop plot. Default NA, 1 is used.
AA_start - 1 is used as the xlim start.}

\item{AA_end}{The ending residue position for the lollipop plot. Default NA, max(AA_POS) is used.
AA_end + 1 is used as the xlim end.}

\item{return_individual_plots}{If TRUE, individual tracks from the lollipop plot is return. This is useful
when aligning extract track to the plot.}
}
\value{
lollipop plot
}
\description{
This function graphs a lollipop plot to compare mutational profile between cases and
controls in a given gene. The center ET track is colored as prismatic style, with the most important
ET positions as red. The height of the lollipops reflects allele count. The color and/or size of the
circles reflects EA scores for the mutations. This function fetch pre stored ET scores, which only
works for human and E. coli reference proteins. See LollipopPlot for ploting mutations only from the
case.
}
\examples{
# Prepare variant data
set.seed(940)
mut_case <- read_tsv(file.path(system.file("extdata", package = "EvoTrace"),
                              "basS_muts.tsv"),
                     show_col_types = FALSE) \%>\%
  mutate(AC = runif(11, 0, 1000)) \%>\%
  mutate(AC = round(AC)) \%>\%
  mutate(AF = AC/1000)

mut_ctrl <- mut_case \%>\%
  mutate(AC = runif(11, 0, 1000)) \%>\%
  mutate(AC = round(AC)) \%>\%
  mutate(AF = AC/1000)

# Make log scale plot using prismatic coloring
LollipopPlot2(variants_case = mut_case, variants_ctrl = mut_ctrl,
              prot_id = "b4112", AC_scale = "log",
              plot_domain = TRUE, show_EA_bin = TRUE,
              EA_color = "prismatic")

# Make linear scale plot using EA_bin coloring
LollipopPlot2(variants_case = mut_case, variants_ctrl = mut_ctrl,
              prot_id = "b4112", AC_scale = "linear",
              plot_domain = FALSE, show_EA_bin = TRUE,
              EA_color = "EA_bin", pad_ratio = 0)
}
