% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PymolFunctions.R
\name{PymolColorChainByET}
\alias{PymolColorChainByET}
\title{Pymol color residues in a given chain using ET}
\usage{
PymolColorChainByET(
  chain,
  position,
  coverage,
  color_type,
  object = NULL,
  coverage_cutoff = NULL
)
}
\arguments{
\item{chain}{The chain in the pymol session that needs to be colored. Input can
be a vector with all the chains that need to be colored.}

\item{coverage}{A double vector for residue ET coverage, range (0,1]}

\item{color_type}{The type of color range to use. Available selections: "ET",
"red_white", "red_white_blue", "white_red", "white_blue, "alphafold", "gray_sacle", "EA_bin".}

\item{object}{A given object in the pymol session that needs to be colored. If is
NULL, then all object in the pymol session will be colored.}

\item{coverage_cutoff}{Only residue below this cutoff will be colored.}

\item{poistion}{A numeric vector specifying the residues that needs to be
colored}
}
\value{
A string vector with pymol commands
}
\description{
Produce pymol commands to color residues in given chains using ET coverage. This function should
be used in combination with PymolLoadFile and PymolExecuteAndSave. Combine PymolLoadFile and
all other pymol cmds with c(), then pipe into PymolExecuteAndSave.
}
