% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PymolFunctions.R
\name{PymolColorChainByResidue}
\alias{PymolColorChainByResidue}
\title{Pymol color residues in a given chain}
\usage{
PymolColorChainByResidue(chain, position, color, object = NULL)
}
\arguments{
\item{chain}{The chain in the pymol session that needs to be colored}

\item{color}{A string vector with hex colors specifying the color. Should be
equal length of position. Color format "0xFFFFFF".}

\item{object}{A given object in the pymol session that needs to be colored. If is
NULL, then all object in the pymol session will be colored.}

\item{poistion}{A numeric vector specifying the residues that needs to be
colored}
}
\value{
A string vector with pymol commands
}
\description{
Produce pymol commands to color residues in a given chain. This function should
be used in combination with PymolLoadFile and PymolExecuteAndSave. Combine PymolLoadFile and
all other pymol cmds with c(), then pipe into PymolExecuteAndSave.
}
