% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PymolFunctions.R
\name{PymolColorChainSingleColor}
\alias{PymolColorChainSingleColor}
\title{Pymol color residues in a given chain with one color}
\usage{
PymolColorChainSingleColor(chain = NULL, color = "white")
}
\arguments{
\item{chain}{The chain in the pymol session that needs to be colored}

\item{color}{A string vector with hex colors specifying the color. Should be
equal length of position. Color format "0xFFFFFF" or color name. Default is white.}
}
\value{
A string vector with pymol commands
}
\description{
Produce pymol commands to color a given chain with on color. This function should
be used in combination with PymolLoadFile and PymolExecuteAndSave. Combine PymolLoadFile and
all other pymol cmds with c(), then pipe into PymolExecuteAndSave.
}
