% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PymolFunctions.R
\name{PymolColorPairsByCovET}
\alias{PymolColorPairsByCovET}
\title{Pymol color residue interaction using CovET}
\usage{
PymolColorPairsByCovET(
  chain1,
  position1,
  chain2,
  position2,
  pair_coverage,
  color_type = c("ET", "red_white", "red_white_blue"),
  group_name = "pairs",
  pair_coverage_cutoff = NULL
)
}
\arguments{
\item{chain1}{Chain ID for the first set of residues}

\item{chain2}{Chain ID for the first set of residues}

\item{color_type}{The type of color range to use. Available selections: "ET",
"red_white", "red_white_blue", "white_red", "white_blue, "alphafold", "gray_sacle", "EA_bin".}

\item{group_name}{Pymol group name to combine all the interactions.}

\item{pair_coverage_cutoff}{Only residue pairs below this cutoff will be shown.}

\item{poistion1}{A numeric vector specifying one side of the interacting
residues}

\item{poistion2}{A numeric vector specifying the other side of the interacting
residues}

\item{coverage}{A double vector for residue pair CovET coverage, range (0,1]}
}
\value{
A string vector with pymol commands
}
\description{
Produce pymol commands to color residues in given chains using ET coverage. This function should
be used in combination with PymolLoadFile and PymolExecuteAndSave. Combine PymolLoadFile and
all other pymol cmds with c(), then pipe into PymolExecuteAndSave.
}
