% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PymolFunctions.R
\name{PymolExecuteAndSave}
\alias{PymolExecuteAndSave}
\title{Execute pymol commands}
\usage{
PymolExecuteAndSave(
  pymol_cmd,
  output_file = NULL,
  pymol_cmd_output = NULL,
  pymol_path = NULL
)
}
\arguments{
\item{pymol_cmd}{a string vector of all pymol command that need to be run}

\item{output_file}{file name of the pymol session to be saved. If is null, pymol
commands will be run, but there is no additional save step.}

\item{pymol_cmd_output}{file name of the pymol script, should end with .pml. Default is
null, the pymol script will be saved as a temporary file, and will be removed after the
run.}

\item{pymol_path}{path to pymol. Default is null, default pymol is called.}
}
\value{
Nother is returned.
}
\description{
pymol commands in pymol_cmd are written as a pymol script, then pymol is evoked
to run that script. Pymol session and pymol script can be saved.
}
