% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PymolFunctions.R
\name{PymolLoadFile}
\alias{PymolLoadFile}
\title{Load pdb/pse file}
\usage{
PymolLoadFile(input_file)
}
\arguments{
\item{input_file}{Path to the pdb/pse file of interest}
}
\value{
A string vector with pymol commands
}
\description{
Produce pymol commands open a pdb/pse file. This function should be used in
combination with PymolExecuteAndSave and other pymol functions. Combine PymolLoadFile and
all other pymol cmds with c(), then pipe into PymolExecuteAndSave.
}
