% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PymolFunctions.R
\name{PymolSaveScene}
\alias{PymolSaveScene}
\title{Pymol save scene}
\usage{
PymolSaveScene(
  scene_name,
  color = 1,
  view = 0,
  active = 0,
  rep = 0,
  frame = 0,
  animate = 0
)
}
\arguments{
\item{scene_name}{Name of the scene in pymol}

\item{color}{0 or 1. Whether color in the current pymol session is saved in the scene.}

\item{view}{0 or 1. Whether view in the current pymol session is saved in the scene.}

\item{active}{0 or 1. Whether active objects in the current pymol session are saved in
the scene.}

\item{rep}{0 or 1. Whether representations in the current pymol session are saved in
the scene.}

\item{frame}{0 or 1. Whether frame in the current pymol session is saved in the scene.
Not sure what this does in pymol, probably relate to movie.}

\item{animate}{0 or 1. Whether animation in the current pymol session is saved in the scene.
Not sure what this does in pymol, probably relate to movie.}
}
\value{
A string vector with pymol commands
}
\description{
Produce pymol commands to save current state of pymol objects as a scene. This
is very useful to store different coloring of an object in pymol.This function should be
used in combination with PymolLoadFile and PymolExecuteAndSave. Combine PymolLoadFile and
all other pymol cmds with c(), then pipe into PymolExecuteAndSave.
}
