% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PymolFunctions.R
\name{PymolSelectResidues}
\alias{PymolSelectResidues}
\title{Pymol select residues}
\usage{
PymolSelectResidues(chain, position, selection_name)
}
\arguments{
\item{chain}{The chain in the pymol session that needs to be selected}

\item{poistion}{A numeric vector specifying the residues that needs to be
selected}
}
\value{
A string vector with pymol commands
}
\description{
Produce pymol commands to select residues in a given chain as a selection.
This function should be used in combination with PymolLoadFile and PymolExecuteAndSave.
Combine PymolLoadFile and all other pymol cmds with c(), then pipe into PymolExecuteAndSave.
}
