% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles.R
\name{ReadClustalMSA}
\alias{ReadClustalMSA}
\title{Read Clustal MSA file}
\usage{
ReadClustalMSA(clustal_path)
}
\arguments{
\item{clustal_path}{Path to MSA file}
}
\value{
A tibble format of the MSA. Each sequence is in one column, each cell is a residue.
The position of each residue in each sequence is also returned. NA is returned in the position
columns for gaps.
}
\description{
Convert Clustal MSA file into a tibble
}
