% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles.R
\name{ReadMSF}
\alias{ReadMSF}
\title{Read MSF multiple sequence alignment file}
\usage{
ReadMSF(msf_file, output_type = c("vector", "df"))
}
\arguments{
\item{msf_file}{Path to MSF file}

\item{output_type}{vector or df}
}
\value{
If vector is returned, then all sequences are stored in a vector, ids are
stored as the names. If a dataframe is returned, then ids are stored in id column and
sequences are stored in seq column.
}
\description{
Read a GCG MSF multiple sequence alignment file. Sequence blocks are
concatenated by ID and periods used as gap characters are converted to hyphens.
}
