% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RewritePDB.R
\name{RewritePDB}
\alias{RewritePDB}
\title{Rewrite PDB file according to linear sequence}
\usage{
RewritePDB(
  pdb_file,
  chains,
  linear_seq,
  fix_positions = TRUE,
  keep_other_chains = TRUE,
  output_file
)
}
\arguments{
\item{pdb_file}{path to PDB file}

\item{chains}{A character vector. Chains that need to be rewritten.}

\item{linear_seq}{A character. The linear sequence.}

\item{fix_positions}{TRUE or FALSE. If TRUE, the positions in the PDB file are repleaced
with linear sequence positions.}

\item{keep_other_chains}{TRUE or FALSE. Whether other chains are kept.}

\item{output_file}{path to output PDB file}
}
\value{
a message reporting the output file is written
}
\description{
The PDB files are often messy. The sequences positions might not match the linear
sequence positions. The pdb file might contain insertions, like His tag. This function reads the
sequence in PDB file and aligns it with the linear sequence. Any positions in the pdb file that
don't align to the linear sequence will be removed. The position number can be rewritten to match
linear positions. Water molecules are also removed. Only coordinates fields are kept in the output
pdb file.
}
