% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectColor.R
\name{SelectColor}
\alias{SelectColor}
\title{Return a preset color range (hex code) with 100 colors}
\usage{
SelectColor(
  color_type = c("ET", "red_white", "red_white_blue", "white_red", "white_blue",
    "alphafold", "gray_scale", "EA_bin", "rev_EA_bin", "red_yellow_green"),
  prefix = "#"
)
}
\arguments{
\item{color_type}{The type of color range to use. Available selections: "ET",
"red_white", "red_white_blue", "white_red", "white_blue, "alphafold", "gray_sacle", "EA_bin",
"rev_EA_bin", "red_yellow_green".}

\item{prefix}{"#" or "0x". Use "0x" when passing color values to pymol.}
}
\value{
a character string with length of 100. Contains hex code of 100 colors.
}
\description{
Return a color scheme of 100 colors. For ET color, the smaller the
index, the warmer the color. For alphafold color, the larger the index, the more
confident of the structure prediction.
}
