% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB3to1.R
\name{SUB3to1}
\alias{SUB3to1}
\title{Convert AA substitutions from 3 letter to 1 letter}
\usage{
SUB3to1(SUB)
}
\arguments{
\item{SUB}{substitution using 3 letter representation}
}
\value{
substitution using 1 letter representation
}
\description{
Convert amino acids from 3 letter to 1 letter in a substitution
}
