% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color_Variants_AlphaFold.R
\name{Color_Variants_AlphaFold}
\alias{Color_Variants_AlphaFold}
\title{Map variants and ET scores to AlphaFold structures}
\usage{
Color_Variants_AlphaFold(
  variants_case,
  variants_ctrl = NULL,
  prot_id,
  pse_output
)
}
\arguments{
\item{variants_case}{A dataframe that stores the mutations in the cases in target gene.
The dataframe should contain at least 2 columns: SUB and EA. Single letter animo acid code should
be used in SUB, eg. L10P. EA can be replaced with other values, but have to scale to the
range of (0, 100].}

\item{variants_ctrl}{A dataframe that stores the mutations in the controls in target gene.}

\item{prot_id}{Protein identification. Use full ENSP for human proteins, eg. ENSP00000388638.1.
Use locus tag for E. coli MG1655 proteins, eg. b4112.}

\item{pse_output}{Path to store the pymol session file.}
}
\value{
A tibble that contains the coloring information. The pse file will
be written to the location indicated by pse_output.
}
\description{
Map variants and ET scores to AlphaFold structures. Pymol is required. Check if Pymol
can be started from command line with "pymol". If not, an open source pymol can be installed
through conda or home brew. The reference amino acid in the mutations will be checked against
the ones in the ET dataset. But the EA scores won't be checked.
}
