% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorPDBByET.R
\name{ColorPDBByET}
\alias{ColorPDBByET}
\title{Pymol color pdb structure using ET}
\usage{
ColorPDBByET(
  pdb_file,
  chain,
  ET_format = c("EA", "UET", "ENSP"),
  ET_file,
  color_type = c("ET", "red_white", "red_white_blue", "gray_scale", "rev_EA_bin",
    "red_yellow_green"),
  coverage_cutoff = 1,
  remaining_color = "white",
  additional_pymol_cmds = NULL,
  output_file,
  output_format = c("pml", "pse"),
  pdb_format = c("pdb", "pdbx")
)
}
\arguments{
\item{pdb_file}{Path to pdb or pse file. Can be path to local file or url ("https://files.rcsb.org/view/1xb7.pdb").}

\item{chain}{The chain in the pymol session that needs to be colored. Can provide multiple chains as a vector.}

\item{ET_format}{Specify the ET file format. "UET" for ET server, "EA" for ET in EA output, and "ENSP"
for precalculated ET scores from EA VEP output.}

\item{ET_file}{Path to ET file. If "ENSP" is used in ET_format, provide ENSP id here, including version.}

\item{color_type}{The type of color range to use. Available selections: "ET",
"red_white", "red_white_blue", "gray_scale", "rev_EA_bin", "red_yellow_green".}

\item{coverage_cutoff}{Value between (0,1]. Only residue below this cutoff will be colored.}

\item{remaining_color}{Color for the remaining chain. Default "white". If set to NULL, color
are not adjusted for those chains. To color heterodimer, color the first chain with other_color as "white" and
save as pse. Then color the second chain using the pse file with other_color = NULL to retain the first chain color.}

\item{additional_pymol_cmds}{Additional pymol commands to append to the pymol script. Store them in a vector.}

\item{output_file}{Path to write the output pymol file.}

\item{output_format}{Output pymol script (pml) or pymol session (pse). Pse output requires pymol to
be accessible from terminal with "pymol".}

\item{pdb_format}{Use "pdb" if the input is traditional pdb. Use "pdbx" for PDBx/mmCIF.}

\item{pml_output}{Path to store the pymol command file.}
}
\value{
A tibble with the sequence alignment between the linear sequence in ET file and the sequence in
the target chain in PDB.
}
\description{
The function first align the linear sequence from the ET file to the sequence in PDB. PDB
structure is then colored according the ET scores. A pymol output file is written to provided path.
The function returns an alignment table. The percent matching between sequence in the ET file and PDB
is printed in the console.
}
\examples{
ColorPDBByET(pdb_file = "https://files.rcsb.org/view/2pjl.pdb", chain = "A", ET_format = "ENSP",
             ET_file = "ENSP00000000442.6", color_type = "ET", coverage_cutoff = 0.3,
             output_format = "pml", output_file = "test.pml")
}
