% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCW_z.R
\name{ET_SCWzscore}
\alias{ET_SCWzscore}
\title{Compute SCW z scores for ET file}
\usage{
ET_SCWzscore(
  pdb_file = NULL,
  chain = NULL,
  SCW_background = NULL,
  ET_file,
  coverage = 0.3,
  adjust_gap = FALSE
)
}
\arguments{
\item{pdb_file}{Path to pdb file, needs to be specified if SCW_bg is not provided}

\item{chain}{Chain in the pdb file to use, needs to be specified if SCW_bg is not provided}

\item{SCW_background}{A list with parameters required for SCW z score calculation.
This is obtained through GetSCWBackground function. Needs to be specified if pdb_file/chain
are not provided}

\item{ET_file}{Path to ranks file}

\item{coverage}{A numeric (0-1) value to specify the top fraction of ET residues are used
in SCW z score calculation}

\item{adjust_gap}{Logic. When False, this function behaves like PyET viewer, all residues
in the pdb structure are used to calculate SCW background. When TRUE, only aligned positions
in the pdb structure are used to calculate this background. Should set to TRUE when there is
a large insertion in the pdb comparing to the ET sequence. Adjust_gap only works when using
pdb_file/chain as input.}
}
\value{
a df with biased and unbiased SCW z scores.
}
\description{
This function calculates the biased and unbiased SCW z scores
for top residues in the ET ranks file. Can use pdb_file/chain or pre calculated SCW_background
as input.
}
