% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetColor.R
\name{GetColor}
\alias{GetColor}
\title{Assign colors for a given numeric vector}
\usage{
GetColor(
  x,
  lower_bound,
  upper_bound,
  color_type = c("ET", "red_white", "red_white_blue", "white_red", "white_blue",
    "alphafold", "gray_scale", "EA_bin")
)
}
\arguments{
\item{x}{A numeric vector}

\item{lower_bound}{The lower bound of the scale}

\item{upper_bound}{The upper bound of the scale}

\item{color_type}{The color ranges to choose from.}
}
\value{
A vector of hex colors
}
\description{
There are 100 colors stored in each color_type.
The color value assigned for each element in x is based on its relative position (percent ranking)
regarding the lower and upper bounds. The assignment is upper bound included. A value equal the lower
bound will not be assigned correctly, and an error will be given.
}
