% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCW_z.R
\name{GetSCWBackgound}
\alias{GetSCWBackgound}
\title{Compute the background stats for SCW z score calculation}
\usage{
GetSCWBackgound(pdb_file, chain, dist_cutoff = 4, resi = NULL)
}
\arguments{
\item{pdb_file}{path to pdb file}

\item{chain}{the protein chain SCW z score is calculated on}

\item{dist_cutoff}{the distance cutoff that consider two atoms are adjacent}

\item{resi}{a vector of residues numbers that are included in SCW bg calculation.
If NULL, then all residues in that chain is used.}
}
\value{
a list with parameters that are needed for SCW z scores calculation
}
\description{
Calculates the required stats for w_hat, w^2_hat calculation in
SCW z scores. Residues are consider adjacent when any of their heavy atoms are
within the cutoff range. The output is feed into ComputeSCWzscore function for
z score calculation.
}
