% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCW_z.R
\name{GetSCWBackgound_reduce_memory}
\alias{GetSCWBackgound_reduce_memory}
\title{Compute the background stats for SCW z score calculation}
\usage{
GetSCWBackgound_reduce_memory(
  pdb_file,
  chain,
  dist_cutoff = 4,
  resi = NULL,
  temp_dir = "temp",
  cores = 1
)
}
\arguments{
\item{pdb_file}{path to pdb file}

\item{chain}{the protein chain SCW z score is calculated on}

\item{dist_cutoff}{the distance cutoff that consider two atoms are adjacent}

\item{resi}{a vector of residues numbers that are included in SCW bg calculation.
If NULL, then all residues in that chain is used.}

\item{temp_dir}{temp dir to store intermediate files. Default is to create a temp dir
in the current working directory.}

\item{cores}{number of cores to parallel computing.}
}
\value{
a list with parameters that are needed for SCW z scores calculation
}
\description{
Calculates the required stats for w_hat, w^2_hat calculation in
SCW z scores. Residues are consider adjacent when any of their heavy atoms are
within the cutoff range. Intermediate files are produced to resolve memory issue.
The output is feed into ComputeSCWzscore function for
z score calculation.
}
