% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles.R
\name{ReadET}
\alias{ReadET}
\title{Read legacy ET format}
\usage{
ReadET(ET_file)
}
\arguments{
\item{ET_file}{Path to ET file}
}
\value{
a dataframe(tibble) with all the columns from ET output
}
\description{
Read legacy ET file and output a tibble
}
