% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFiles.R
\name{ReadFasta}
\alias{ReadFasta}
\title{Read fasta file}
\usage{
ReadFasta(fasta_file, output_type = c("vector", "df"))
}
\arguments{
\item{fasta_file}{Path to fasta file}

\item{output_type}{vector or df}
}
\value{
If vector is returned, then all sequences are stored in a vector, ids are
stored as the names. If a dataframe is returned, then ids are stored in id column and
sequences are stored in seq column.
}
\description{
Read fasta file
}
